+++ core/src/main/kotlin/org/evomaster/core/search/gene/ImmutableDataHolderGene.kt
@@ -48,5 +48,5 @@ class ImmutableDataHolderGene(name: String, val value: String, val inQuotes: Boo
         return value == other.value
     }
 
-    override fun mutationWeight(): Int = 0
+    override fun mutationWeight(): Double = 0.0
 }
